pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.4",
    "llama-index>=0.14.10",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.5",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "model2vec>=0.7.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "ruff>=0.14.9",
]
